refactor: tx simulation hooks and types
