+++ suite-common/connect-popup/src/index.ts
@@ -4,3 +4,4 @@ export * from './connectPopupMiddleware';
 export * from './connectPopupReducer';
 export * from './connectPopupTypes';
 export * from './connectPopupPromiseManager';
+export * from './hooks/useTxSimulationPopupCall';
