+++ suite-common/tx-simulation/package.json
@@ -11,10 +11,9 @@
     },
     "dependencies": {
         "@blockaid/client": "^0.56.0",
-        "@suite-common/connect-popup": "workspace:^",
         "@suite-common/react-query": "workspace:^",
         "@suite-common/wallet-config": "workspace:^",
-        "@trezor/connect": "workspace:^",
+        "@suite-common/wallet-types": "workspace:*",
         "react": "18.2.0"
     }
 }
