+++ suite-common/tx-simulation/src/index.ts
@@ -1,4 +1,7 @@
 export * from './client';
-export * from './hooks';
 export * from './types';
-export * from './utils';
+export { getSimulationErrorRiskLevel } from './utils';
+export { useDappScan } from './hooks/useDappScan';
+export { type TxSimulationEVMResult } from './hooks/useTxSimulationEVM';
+export { useTxSimulation } from './hooks/useTxSimulation';
+export type { AccountSummary, TransactionSimulation } from '@blockaid/client/resources/evm';
