--- suite-common/tx-simulation/src/utils.ts
@@ -1,7 +1,0 @@
-export const getSimulationErrorRiskLevel = (message: string) => {
-    if (message.includes('Unsupported EIP-712 message type')) {
-        return 'warning';
-    }
-
-    return 'error';
-};
