+++ suite-common/tx-simulation/tsconfig.json
@@ -2,10 +2,9 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../connect-popup" },
         { "path": "../react-query" },
         { "path": "../wallet-config" },
-        { "path": "../../packages/connect" }
+        { "path": "../wallet-types" }
     ],
     "include": [".", "**/*.json"]
 }
